Fix bad arguments passing
Explicitly pass a new object with the parsed arguments, rather than passing the raw
Minimist object.

---

Some time ago, there seems to have been a change in Minimist that disallowed adding new
keys, resulting in passing no `files` array to the server, causing it to crash, which
caused `stylelint_d` to simply show an error. To fix this, an object that is not the
Minimist instance is passed with the expected arguments.

This bug (which seems to have existed for a while) underscores the need for tests,
specifically integration style tests.